+++ SteakBot.Core/SteakBot.Core.csproj
@@ -206,7 +206,7 @@
   <ItemGroup>
     <Compile Include="Bot.cs" />
     <Compile Include="EventHandlers\LogHandler.cs" />
-    <Compile Include="EventHandlers\MessageReceivedHandler.cs" />
+    <Compile Include="EventHandlers\CommandMessageEventHandler.cs" />
     <Compile Include="EventHandlers\ReactionEventHandler.cs" />
     <Compile Include="EventHandlers\VoiceStateEventHandler.cs" />
     <Compile Include="ManualCommandHandler.cs" />
